Move repo init helper to git module
